Move and rename directories for repository layer
