+++ src/controllers/ContactController.ts
@@ -1,6 +1,6 @@
 import {Request, Response} from 'express';
 
-import models from '../services/models';
+import models from '../db/repository';
 const {
   Contact,
 } = models;
