+++ src/db/repository/Configuration.ts
@@ -1,4 +1,4 @@
-import models from '../../db/models';
+import models from '../models';
 import logger from '../../logs/logger';
 
 import LogDatabase from '../../interfaces/typeDefinitions/LogDatabase';
