+++ src/db/repository/Contact.ts
@@ -1,4 +1,4 @@
-import models from '../../db/models';
+import models from '../models';
 import logger from '../../logs/logger';
 
 import ContactInterface from '../../interfaces/typeDefinitions/Contact';
