+++ src/db/repository/TweetQueue.ts
@@ -1,4 +1,4 @@
-import models from '../../db/models';
+import models from '../models';
 import {Op} from 'sequelize';
 import logger from '../../logs/logger';
